# Much faster implementation