--- a/index.browser.js
+++ b/index.browser.js
@@ -1,6 +1,6 @@
 // This file replaces `index.js` in bundlers like webpack or Rollup,
 // according to `browser` config in `package.json`.
-
+import { urlAlphabet } from './url-alphabet/index.js';
 export { urlAlphabet } from './url-alphabet/index.js'
 
 export let random = bytes => crypto.getRandomValues(new Uint8Array(bytes))
@@ -49,11 +49,10 @@ export let customAlphabet = (alphabet, size = 21) =>
 
 export let nanoid = (len = 21) => {
   let id = "";
-  let chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_";
   let rand = crypto.getRandomValues(new Uint8Array(len));
   // Using the bitwise AND operator to "cap" the value of
   // the random byte from 255 to 63, in that way we can make sure
   // that the value will be a valid index for the "chars" string.
-  for (let i = 0; i < len; i++) id += chars[rand[i] & 63];
+  for (let i = 0; i < len; i++) id += urlAlphabet[rand[i] & 63];
   return id;
 }