@ai 2024-01-31
1. It is the file generated from `index.browser.js`. You need to fix it first.
2. But browser's file should be tested in browsers. For Node/Bun we have another file `index.js` which has different optimizations.