@samuelgozi 2024-01-31
Tested on latest version:

Chrome
```
Old 1634ms
New 1245ms
```

Safari
```
Old 1208ms
New 760ms
```

Firefox
```
Old 1207
New 742
```

Browser support doesn't change much.
I will update the source file.