@samuelgozi 2024-01-31
Ok i will fix everything you mentioned.
In regards to the use of the string inside the function, I don't think we should hoist it manually since I didn't see any effect on performance, but I do think that this way it looks nicer.

Maybe I should remove the comment instead since it not important