@ai 2024-01-31
OK, let's keep string as is. I will do some optimizations later for smaller bundle size.