@samuelgozi 2024-01-31
Done. Package size limit has exceeded by 34 B