@samuelgozi 2024-02-01
`Package size limit has exceeded by 15 B`
TIL, very cool!

Made the changes