@ai 2024-02-01
I also reduced 8 bytes by reusing `size` variable 4db3464