@samuelgozi 2024-02-04
Using the while loop is actually a little bit faster.
Very nice idea!